Fixed javadoc errors.
If you clone the nailgun repository, it builds just fine. However, if
you try to use this version of nailgun with scalafmt [1], it doesn't
actually work [2]. In [2], they suggest building the nailgun-all-0.9.1
tag. But, this tag doesn't build because of javadoc errors. This commit
fixes those errors so that the code builds again.

[1]: https://scalameta.org/scalafmt/#Nailgun
[2]: https://github.com/scalameta/scalafmt/issues/1115